wip: rainbow lammps simulation
Problem: we need a more controlled experiment for compatibility metadata.
Using the previous lammps images with a small set of clusters, and reducing
to that based on actually building them, is limiting.
Solution: Instead, generate a set of 48 clusters so that each jobspec (the
previous ones for lammps, varying in os name and version, mpi implementation,
has/does not have gpu, and architecture) is matched to a perfect cluster.
We can then programmatically generate the cluster nodes AND subsystems,
and run everything in a simulation locally, where we actually do
register the 48 clusters to rainbow, and then add subsystems one at a time
and "run" the jobspecs to see the matchings. Likely we could even vary
the ordering of the subsystem addition to get a sense for the value added
for attributes, but I think I will hold off on that. Also note that the
current subsystem metadata is completely flat - meaning that a node in
a subsystem graph is a complete attribute-key=value. This is by no means
intended to be something final, just as a match to the first simplest
algorithm of "do an exact match of the node type" so thus the attributes
are embedded in the type.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>