Add reliabuild version range experiments (#1)
* wip: running reliabuild experiments

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* refactor: match based on versions

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* wip: refactored reliabuild experiments

This experiment has become primarily about the build use case,
and specifically package versions as the compatibility metadata.
Since we cannot generate every perfect cluster, I think the result
is going to be an example of how adding too much (in terms of
requirements) leads to a poorer outcomes. Of course it is entirely
based on the number of clusters I made, etc. I currently have only
100 clusters for about 10K jobs and over 200 dependency metadata
(of course not every one is relevant for every package) so the odds
of getting a matching cluster are pretty slim when you start asking
for more detail. Hopefully folks can help to think of smarter
experiments too.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* results: add raw files in case I do something stupid

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

* add results

Signed-off-by: vsoch <vsoch@users.noreply.github.com>

---------

Signed-off-by: vsoch <vsoch@users.noreply.github.com>
Co-authored-by: vsoch <vsoch@users.noreply.github.com>